reorient mongo client types
